идентификация с тестами

+++ src/main/java/com/amr/project/model/dto/PersonalDataDto.java
@@ -22,7 +22,7 @@ import java.util.List;
 public class PersonalDataDto {
 
     private Long id;
-    private int passport;
+    private Long passport;
     private Date dateOfIssue;
     private String authority;
     private String placeOfBirth;
